+++ simulation/traffic_simulator/include/traffic_simulator/entity/entity_base.hpp
@@ -38,8 +38,6 @@ namespace traffic_simulator
 {
 namespace entity
 {
-enum class Direction { STRAIGHT = 0, LEFT = 1, RIGHT = 2 };
-
 class EntityBase
 {
 public:
@@ -121,6 +119,8 @@ public:
 
   virtual void setTargetSpeed(double target_speed, bool continuous) = 0;
 
+  virtual void setTargetSpeed(const RelativeTargetSpeed & target_speed, bool continuous) = 0;
+
   virtual void setTrafficLightManager(
     const std::shared_ptr<traffic_simulator::TrafficLightManagerBase> & ptr)
   {
@@ -151,7 +151,23 @@ public:
     const double target_speed, const SpeedChangeTransition transition,
     const SpeedChangeConstraint constraint, const bool continuous);
 
+  virtual void requestSpeedChange(
+    const RelativeTargetSpeed & target_speed, const SpeedChangeTransition transition,
+    const SpeedChangeConstraint constraint, const bool continuous);
+
   virtual void requestLaneChange(const std::int64_t){};
+
+  virtual void requestLaneChange(const traffic_simulator::lane_change::Parameter &){};
+
+  void requestLaneChange(
+    const traffic_simulator::lane_change::AbsoluteTarget & target,
+    const traffic_simulator::lane_change::TrajectoryShape trajectory_shape,
+    const lane_change::Constraint & constraint);
+
+  void requestLaneChange(
+    const traffic_simulator::lane_change::RelativeTarget & target,
+    const traffic_simulator::lane_change::TrajectoryShape trajectory_shape,
+    const lane_change::Constraint & constraint);
 
   virtual void requestWalkStraight()
   {